+++ packages/dso-toolkit/src/components/document-component/document-component.models.ts
@@ -2,6 +2,7 @@ export interface DocumentComponent<TemplateFnReturnType> {
   alternativeTitle?: string;
   annotated?: boolean;
   bevatOntwerpInformatie?: boolean;
+  content?: string;
   children?: TemplateFnReturnType;
   dsoAnnotationToggle?: (e: CustomEvent<DocumentComponentAnnotationToggleEvent>) => void;
   dsoToggle?: (e: CustomEvent<DocumentComponentToggleEvent>) => void;