+++ packages/dso-toolkit/src/components/ozon-content/ozon-content.stories-of.ts
@@ -5,9 +5,7 @@ import { content } from "./ozon-content.content.js";
 import { OzonContent } from "./ozon-content.models.js";
 
 export interface OzonContentTemplates<TemplateFnReturnType> {
-  ozonContentTemplate: (ozonContentProperties: OzonContent<TemplateFnReturnType>) => TemplateFnReturnType;
-  prefix?: TemplateFnReturnType;
-  suffix?: TemplateFnReturnType;
+  ozonContentTemplate: (ozonContentProperties: OzonContent) => TemplateFnReturnType;
 }
 
 export function storiesOfOzonContent<Implementation, Templates, TemplateFnReturnType>(
@@ -23,8 +21,8 @@ export function storiesOfOzonContent<Implementation, Templates, TemplateFnReturn
       argTypes: ozonContentArgTypes,
     });
 
-    const template = templateMapper<OzonContentArgs>((args, { ozonContentTemplate, prefix, suffix }) =>
-      ozonContentTemplate(ozonContentArgsMapper(args, prefix, suffix))
+    const template = templateMapper<OzonContentArgs>((args, { ozonContentTemplate }) =>
+      ozonContentTemplate(ozonContentArgsMapper(args))
     );
 
     content.forEach((story) => {
@@ -32,10 +30,6 @@ export function storiesOfOzonContent<Implementation, Templates, TemplateFnReturn
         args: componentArgs<Omit<OzonContentArgs, "dsoAnchorClick" | "dsoClick">>({
           content: story.content,
           inline: false,
-          interactive: false,
-          deleted: false,
-          prefix: story.args?.prefix,
-          suffix: story.args?.suffix,
           ...story.args,
         }),
       });