+++ packages/react/src/components/ozon-content/ozon-content.react-template.tsx
@@ -4,18 +4,13 @@ import * as React from "react";
 import { DsoOzonContent } from "../../components";
 import { ComponentImplementation } from "../../templates";
 
-export const reactOzonContent: ComponentImplementation<OzonContent<JSX.Element>> = {
+export const reactOzonContent: ComponentImplementation<OzonContent> = {
   component: "ozonContent",
   implementation: "react",
   template: () =>
-    function ozonContentTemplate({ slotName, content, prefix, suffix, dsoAnchorClick }) {
+    function ozonContentTemplate({ slotName, content, dsoAnchorClick }) {
       const slotProp = slotName && { slot: slotName };
 
-      return (
-        <DsoOzonContent content={content} onDsoAnchorClick={dsoAnchorClick} {...slotProp}>
-          {prefix ? typeof prefix === "string" ? <span slot="prefix">{prefix}</span> : prefix : null}
-          {suffix ? typeof suffix === "string" ? <span slot="suffix">{suffix}</span> : suffix : null}
-        </DsoOzonContent>
-      );
+      return <DsoOzonContent content={content} onDsoAnchorClick={dsoAnchorClick} {...slotProp} />;
     },
 };